requirements.txt
gradio
pillow
git+https://github.com/facebookresearch/segment-anything.git
opencv-python
torch
torchvision
torchaudio
numpy<2.0  # Temporalmente necesario por compatibilidad
#albumentations
albumentations==0.5.2
omegaconf
einops
#it+https://github.com/advimman/lama.git
types-requests
gdown
webdataset
scikit-image
easydict
scikit-learn
pytorch_lightning
kornia
diffusers
transformers
accelerate
safetensors
